Chapter10/graphql/requirements.txt
fastapi
uvicorn
strawberry-graphql[fastapi]